Created useGame Hook to encapsulate websocket, will put chess game stuff in there too.
